[refactor] Moving LinkedIn lib to plugin Moving the library to the only extension that uses it.
Classes were moved to individual files to comply with our style linter.

@TODO: Replace with Composer package?
