+++ core/plugins/authentication/linkedin/simplelinkedin-php/LinkedInException.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * 'LinkedInException' class declaration.
+ *
+ * This class extends the base 'Exception' class.
+ *
+ * @access public
+ * @package classpackage
+ */
+class LinkedInException extends Exception
+{
+}
