+++ core/plugins/authentication/linkedin/simplelinkedin-php/LinkedInOAuthException.php
@@ -1,0 +1,7 @@
+<?php
+/* Generic exception class
+ */
+class LinkedInOAuthException extends Exception
+{
+	// pass
+}
