+++ core/plugins/authentication/linkedin/simplelinkedin-php/OAuth.php
@@ -1,0 +1,12 @@
+<?php
+include_once __DIR__ . '/LinkedInOAuthException.php';
+include_once __DIR__ . '/OAuthConsumer.php';
+include_once __DIR__ . '/OAuthToken.php';
+include_once __DIR__ . '/OAuthSignatureMethod.php';
+include_once __DIR__ . '/OAuthSignatureMethod_HMAC_SHA1.php';
+include_once __DIR__ . '/OAuthSignatureMethod_PLAINTEXT.php';
+include_once __DIR__ . '/OAuthSignatureMethod_RSA_SHA1.php';
+include_once __DIR__ . '/OAuthRequest.php';
+include_once __DIR__ . '/OAuthServer.php';
+include_once __DIR__ . '/OAuthDataStore.php';
+include_once __DIR__ . '/OAuthUtil.php';
