+++ core/plugins/authentication/linkedin/simplelinkedin-php/OAuthConsumer.php
@@ -1,0 +1,19 @@
+<?php
+
+class OAuthConsumer
+{
+	public $key;
+	public $secret;
+
+	public function __construct($key, $secret, $callback_url=null)
+	{
+		$this->key = $key;
+		$this->secret = $secret;
+		$this->callback_url = $callback_url;
+	}
+
+	public function __toString()
+	{
+		return "OAuthConsumer[key=$this->key,secret=$this->secret]";
+	}
+}
